✨ Feature: LD-144 코멘트 반영
- Background 컴포넌트화
- 툴팁 첫 반응 등록 전 까지
- offset 값 변경
등
