+++ Projects/LastDance/LastDance/Resources/Assets.xcassets/Aperture.imageset/Contents.json
@@ -2,20 +2,14 @@
   "images" : [
     {
       "filename" : "Aperture.svg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "idiom" : "universal"
     }
   ],
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "preserves-vector-representation" : true
   }
 }
