+++ Projects/LastDance/LastDance/Resources/Assets.xcassets/bauhausArt08.imageset/Contents.json
@@ -1,21 +1,15 @@
 {
   "images" : [
     {
-      "filename" : "bauhausArt08.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "filename" : "bauhausArt08.svg",
+      "idiom" : "universal"
     }
   ],
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "preserves-vector-representation" : true
   }
 }
