+++ Projects/LastDance/LastDance/Sources/Presentation/Reaction/ViewModel/ReactionInputViewModel.swift
@@ -82,6 +82,12 @@ final class ReactionInputViewModel: ObservableObject {
                     self.message = ""
                     self.selectedCategories.removeAll()
                     Log.debug("반응 저장 성공")
+                    
+                    // 첫 리액션 등록 플래그 저장
+                    if !UserDefaults.standard.bool(forKey: .hasRegisteredFirstReaction) {
+                        UserDefaults.standard.set(true, forKey: .hasRegisteredFirstReaction)
+                    }
+                    
                     completion(true)
                 case .failure(let error):
                     Log.debug("반응 저장 실패: \(error)")
